feat: 🎸 Fix checkout to only have checkout
