+++ apps/hems/src/app/pages/checkout/checkin.component.html
@@ -1,83 +1,6 @@
 <!-- Checkin section -->
 <main>
   <div class="table-actions">
-    <h1>Check In</h1>
-
-    <button
-      mat-raised-button
-      class="action-btn"
-      (click)="openCheckinCreateDialog()"
-    >
-      + Add Check In
-    </button>
-  </div>
-  <hr />
-
-  <table mat-table [dataSource]="checkinLuggage" class="mat-elevation-z8">
-    <ng-container matColumnDef="room">
-      <th mat-header-cell *matHeaderCellDef>Room nr.</th>
-      <td mat-cell *matCellDef="let item">{{ item.room }}</td>
-    </ng-container>
-    <!-- <ng-container matColumnDef="roomReady">
-      <th mat-header-cell *matHeaderCellDef>Room<br />Status</th>
-      <td mat-cell *matCellDef="let item">{{ item.roomReady }}</td>
-    </ng-container> -->
-    <ng-container matColumnDef="name">
-      <th mat-header-cell *matHeaderCellDef>Name</th>
-      <td mat-cell *matCellDef="let item">{{ item.name }}</td>
-    </ng-container>
-    <ng-container matColumnDef="arrivalTime">
-      <th mat-header-cell *matHeaderCellDef>Time of<br />arrival</th>
-      <td mat-cell *matCellDef="let item">
-        {{ item.arrivalTime ? (item.arrivalTime | date: 'HH:mm') : '-' }}
-      </td>
-    </ng-container>
-    <ng-container matColumnDef="bags">
-      <th mat-header-cell *matHeaderCellDef>Nr. of<br />bags</th>
-      <td mat-cell *matCellDef="let item">{{ item.bags }}</td>
-    </ng-container>
-    <ng-container matColumnDef="tagNr">
-      <th mat-header-cell *matHeaderCellDef>Tag nr.</th>
-      <td mat-cell *matCellDef="let item">{{ item.tagNr }}</td>
-    </ng-container>
-    <ng-container matColumnDef="bbLr">
-      <th mat-header-cell *matHeaderCellDef>BB in LR</th>
-      <td mat-cell *matCellDef="let item">{{ item.bbLr }}</td>
-    </ng-container>
-    <ng-container matColumnDef="location">
-      <th mat-header-cell *matHeaderCellDef>Location</th>
-      <td mat-cell *matCellDef="let item">{{ item.location }}</td>
-    </ng-container>
-    <ng-container matColumnDef="completedAt">
-      <th mat-header-cell *matHeaderCellDef>
-        Time in<br />
-        room
-      </th>
-      <td mat-cell *matCellDef="let item">
-        {{ item.completedAt ? (item.completedAt | date: 'HH:mm') : '-' }}
-      </td>
-    </ng-container>
-    <ng-container matColumnDef="bbOut">
-      <th mat-header-cell *matHeaderCellDef>BB up</th>
-      <td mat-cell *matCellDef="let item">{{ item.bbOut ?? '-' }}</td>
-    </ng-container>
-    <ng-container matColumnDef="comments">
-      <th mat-header-cell *matHeaderCellDef>Comments</th>
-      <td mat-cell *matCellDef="let item">{{ item.comments ?? '-' }}</td>
-    </ng-container>
-
-    <tr mat-header-row *matHeaderRowDef="checkinColumns"></tr>
-    <tr
-      mat-row
-      [ngClass]="row.completedAt ? 'completed' : ''"
-      *matRowDef="let row; columns: checkinColumns"
-      (click)="openCheckinEditDialog(row)"
-    ></tr>
-  </table>
-
-  <br />
-
-  <div class="table-actions">
     <h1>Check Out</h1>
 
     <button
