--- apps/hems/src/app/pages/checkout/dialogs/checkinDialogs/createCheckinDialog/create-checkin-dialog.component.html
@@ -1,203 +1,0 @@
-<form
-  (ngSubmit)="onSubmit()"
-  [formGroup]="form"
-  data-cy="create-checkin-dialog"
->
-  <h1>Create check in</h1>
-  <hr />
-
-  <div class="dialog-row">
-    <mat-form-field autocomplete="off" class="input" appearance="fill">
-      <mat-label class="input-label">Room nr.</mat-label>
-      <input
-        #room
-        type="text"
-        data-cy="update-checkin-room-input"
-        formControlName="room"
-        matInput
-        placeholder="123"
-      />
-      <mat-error *ngIf="form.get('room')?.invalid && form.get('room')?.dirty">
-        Must be a number from 100-999
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Room status</mat-label>
-      <mat-select
-        formControlName="roomReady"
-        data-cy="update-checkin-room-status"
-      >
-        <mat-option value="false">Not ready</mat-option>
-        <mat-option value="true">Ready</mat-option>
-      </mat-select>
-    </mat-form-field>
-  </div>
-
-  <div class="dialog-row">
-    <mat-form-field
-      autocomplete="off"
-      class="input full-width"
-      appearance="fill"
-    >
-      <mat-label class="input-label">Name</mat-label>
-      <input
-        #name
-        type="text"
-        data-cy="update-checkin-name"
-        formControlName="name"
-        matInput
-        placeholder="Name"
-      />
-      <mat-error *ngIf="form.get('name')?.invalid && form.get('name')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Time of arrival</mat-label>
-      <input
-        #arrivalTime
-        matInput
-        [ngxMatDatetimePicker]="pickerArrival"
-        placeholder="dd/MM HH:mm"
-        formControlName="arrivalTime"
-      />
-      <mat-datepicker-toggle
-        matSuffix
-        [for]="$any(pickerArrival)"
-      ></mat-datepicker-toggle>
-      <ngx-mat-datetime-picker
-        #pickerArrival
-        [showSpinners]="true"
-        [showSeconds]="false"
-        [touchUi]="true"
-      >
-      </ngx-mat-datetime-picker>
-    </mat-form-field>
-  </div>
-
-  <div class="dialog-row">
-    <mat-form-field
-      autocomplete="off"
-      class="input one-third"
-      appearance="fill"
-    >
-      <mat-label class="input-label">Nr of bags</mat-label>
-      <input
-        #bags
-        type="text"
-        data-cy="update-checkin-room-input"
-        formControlName="bags"
-        matInput
-        placeholder="1+3"
-      />
-      <mat-error *ngIf="form.get('bags')?.invalid && form.get('bags')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field
-      autocomplete="off"
-      class="input one-third"
-      appearance="fill"
-    >
-      <mat-label class="input-label">Tag nr</mat-label>
-      <input
-        #tagNr
-        type="text"
-        data-cy="update-checkin-tag-nr"
-        formControlName="tagNr"
-        matInput
-        placeholder="123"
-      />
-      <mat-error *ngIf="form.get('tagNr')?.invalid && form.get('tagNr')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-  </div>
-
-  <div class="dialog-row">
-    <mat-form-field
-      autocomplete="off"
-      class="input one-third"
-      appearance="fill"
-    >
-      <mat-label class="input-label">BB in LR</mat-label>
-      <input
-        #bbLr
-        type="text"
-        data-cy="update-checkin-name"
-        formControlName="bbLr"
-        matInput
-        placeholder="Initials"
-      />
-      <mat-error *ngIf="form.get('bbLr')?.invalid && form.get('bbLr')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" class="input" appearance="fill">
-      <mat-label class="input-label">Location</mat-label>
-      <input
-        #location
-        type="text"
-        data-cy="update-checkin-location"
-        formControlName="location"
-        matInput
-        placeholder="LR1"
-      />
-      <mat-error
-        *ngIf="form.get('location')?.invalid && form.get('location')?.dirty"
-      >
-        Required
-      </mat-error>
-    </mat-form-field>
-  </div>
-
-  <mat-form-field autocomplete="off" class="full-width" appearance="fill">
-    <mat-label class="input-label">Comments</mat-label>
-    <textarea
-      type="text"
-      data-cy="update-checkin-comments"
-      formControlName="comments"
-      matInput
-      placeholder="Comments"
-    ></textarea>
-    <mat-error
-      *ngIf="form.get('comments')?.invalid && form.get('comments')?.dirty"
-    >
-      Can't be more than 1000 characters
-    </mat-error>
-  </mat-form-field>
-
-  <label class="checkbox-container">
-    The guest has given consent to the processing of personal data
-    <input
-      type="checkbox"
-      [checked]="guestApprovedGDPR"
-      (change)="guestApprovedGDPR = !guestApprovedGDPR"
-    />
-    <span class="checkmark" data-cy="checkbox-banner"></span>
-  </label>
-
-  <div class="dialog-row">
-    <button
-      mat-raised-button
-      class="secondary-btn"
-      data-cy="create-checkin-cancel-btn"
-      mat-dialog-close
-    >
-      Cancel
-    </button>
-
-    <button
-      mat-raised-button
-      class="action-btn"
-      data-cy="create-checkin-create-btn"
-      type="submit"
-    >
-      + Add checkin
-    </button>
-  </div>
-</form>
