--- apps/hems/src/app/pages/checkout/dialogs/checkinDialogs/updateCheckinDialog/update-checkin-dialog.component.html
@@ -1,208 +1,0 @@
-<form
-  (ngSubmit)="onSubmit()"
-  [formGroup]="form"
-  data-cy="update-checkin-dialog"
->
-  <h1>Update Check In</h1>
-  <hr />
-
-  <div class="dialog-row">
-    <mat-form-field autocomplete="off" class="input" appearance="fill">
-      <mat-label class="input-label">Room nr.</mat-label>
-      <input
-        #room
-        type="text"
-        data-cy="update-checkin-room-input"
-        formControlName="room"
-        matInput
-        placeholder="123"
-      />
-      <mat-error *ngIf="form.get('room')?.invalid && form.get('room')?.dirty">
-        Must be a number from 100-999
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Room status</mat-label>
-      <mat-select
-        formControlName="roomReady"
-        data-cy="update-checkin-room-status"
-      >
-        <mat-option value="false">Not ready</mat-option>
-        <mat-option value="true">Ready</mat-option>
-      </mat-select>
-    </mat-form-field>
-  </div>
-
-  <div class="dialog-row">
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Name</mat-label>
-      <input
-        #name
-        type="text"
-        data-cy="update-checkin-name"
-        formControlName="name"
-        matInput
-        placeholder="John Doe"
-      />
-      <mat-error *ngIf="form.get('name')?.invalid && form.get('name')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Time of arrival</mat-label>
-      <input
-        #arrivalTime
-        matInput
-        [ngxMatDatetimePicker]="pickerArrival"
-        placeholder="dd/MM HH:mm"
-        formControlName="arrivalTime"
-      />
-      <mat-datepicker-toggle
-        matSuffix
-        [for]="$any(pickerArrival)"
-      ></mat-datepicker-toggle>
-      <ngx-mat-datetime-picker
-        #pickerArrival
-        [showSpinners]="true"
-        [showSeconds]="false"
-        [touchUi]="true"
-      >
-      </ngx-mat-datetime-picker>
-    </mat-form-field>
-  </div>
-
-  <div class="dialog-row">
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Nr of bags</mat-label>
-      <input
-        #bags
-        type="text"
-        data-cy="update-checkin-room-input"
-        formControlName="bags"
-        matInput
-        placeholder="1+3"
-      />
-      <mat-error *ngIf="form.get('bags')?.invalid && form.get('bags')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Tag nr.</mat-label>
-      <input
-        #tagNr
-        type="text"
-        data-cy="update-checkin-tag-nr"
-        formControlName="tagNr"
-        matInput
-        placeholder="123"
-      />
-      <mat-error *ngIf="form.get('tagNr')?.invalid && form.get('tagNr')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-  </div>
-
-  <div class="dialog-row">
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>BB in LR</mat-label>
-      <input
-        #bbLr
-        type="text"
-        data-cy="update-checkin-name"
-        formControlName="bbLr"
-        matInput
-        placeholder="JD"
-      />
-      <mat-error *ngIf="form.get('bbLr')?.invalid && form.get('bbLr')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>Location</mat-label>
-      <input
-        #location
-        type="text"
-        data-cy="update-checkin-location"
-        formControlName="location"
-        matInput
-        placeholder="LR1"
-      />
-      <mat-error
-        *ngIf="form.get('location')?.invalid && form.get('location')?.dirty"
-      >
-        Required
-      </mat-error>
-    </mat-form-field>
-  </div>
-
-  <div class="dialog-row">
-    <mat-form-field autocomplete="off" appearance="fill">
-      <mat-label>In room time</mat-label>
-      <input
-        matInput
-        [ngxMatDatetimePicker]="pickerCompletedAt"
-        placeholder="dd/MM HH:mm"
-        formControlName="completedAt"
-      />
-      <mat-datepicker-toggle
-        matSuffix
-        [for]="$any(pickerCompletedAt)"
-      ></mat-datepicker-toggle>
-      <ngx-mat-datetime-picker
-        #pickerCompletedAt
-        [showSpinners]="true"
-        [showSeconds]="false"
-        [touchUi]="true"
-      >
-      </ngx-mat-datetime-picker>
-    </mat-form-field>
-
-    <mat-form-field autocomplete="off" class="input" appearance="fill">
-      <mat-label>BB up</mat-label>
-      <input
-        type="text"
-        data-cy="update-checkin-name"
-        formControlName="bbUp"
-        matInput
-        placeholder="JD"
-      />
-      <mat-error *ngIf="form.get('bbUp')?.invalid && form.get('bbUp')?.dirty">
-        Required
-      </mat-error>
-    </mat-form-field>
-  </div>
-
-  <mat-form-field autocomplete="off" appearance="fill" class="full-width">
-    <mat-label>Comments</mat-label>
-    <input
-      type="text"
-      data-cy="update-checkin-comments"
-      formControlName="comments"
-      matInput
-      placeholder="Comments on the check-in"
-    />
-  </mat-form-field>
-
-  <div class="dialog-row">
-    <button
-      data-cy="update-checkin-cancel-btn"
-      class="secondary-btn"
-      mat-raised-button
-      mat-dialog-close
-    >
-      Cancel
-    </button>
-    <button
-      mat-raised-button
-      data-cy="update-checkin-create-btn"
-      type="submit"
-      class="action-btn"
-    >
-      Update
-    </button>
-  </div>
-</form>
