--- apps/hems/src/app/pages/checkout/dialogs/checkoutDialogs/createCheckoutDialog/create-checkout-dialog.component.ts
@@ -1,118 +1,0 @@
-import { HttpErrorResponse } from '@angular/common/http';
-import { Component, ElementRef, ViewChild } from '@angular/core';
-import {
-  UntypedFormControl,
-  UntypedFormGroup,
-  Validators,
-} from '@angular/forms';
-import { MatDialog } from '@angular/material/dialog';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { LuggageType } from '@hems/interfaces';
-import { LuggageService } from '../../../services/luggage.service';
-
-@Component({
-  selector: 'hems-create-checkout-dialog',
-  templateUrl: './create-checkout-dialog.component.html',
-  styleUrls: [
-    '../../../../../../assets/styles/checkbox.scss',
-    '../../../../../../assets/styles/dialog.scss',
-  ],
-})
-export class CreateCheckoutDialogComponent {
-  form: UntypedFormGroup;
-  checked = true;
-  isLoading = false;
-  guestApprovedGDPR = false;
-
-  @ViewChild('room') roomInput!: ElementRef;
-  @ViewChild('name') nameInput!: ElementRef;
-  @ViewChild('bags') bagsInput!: ElementRef;
-  @ViewChild('tagNr') tagNrInput!: ElementRef;
-  @ViewChild('bbLr') bbLrInput!: ElementRef;
-  @ViewChild('bbDown') bbDownInput!: ElementRef;
-  @ViewChild('location') locationInput!: ElementRef;
-
-  constructor(
-    private service: LuggageService,
-    private snackbar: MatSnackBar,
-    private dialog: MatDialog
-  ) {
-    this.form = new UntypedFormGroup({
-      room: new UntypedFormControl('', [Validators.required]),
-      name: new UntypedFormControl('', [Validators.required]),
-      bags: new UntypedFormControl('', [Validators.required]),
-      tagNr: new UntypedFormControl('', [Validators.required]),
-      bbLr: new UntypedFormControl('', [Validators.required]),
-      bbDown: new UntypedFormControl('', [Validators.required]),
-      location: new UntypedFormControl('', [Validators.required]),
-      comments: new UntypedFormControl('', []),
-    });
-  }
-
-  onSubmit(): void {
-    if (!this.form.valid) {
-      if (this.form.get('room')?.invalid) {
-        this.roomInput.nativeElement.focus();
-      } else if (this.form.get('name')?.invalid) {
-        this.nameInput.nativeElement.focus();
-      } else if (this.form.get('bags')?.invalid) {
-        this.bagsInput.nativeElement.focus();
-      } else if (this.form.get('tagNr')?.invalid) {
-        this.tagNrInput.nativeElement.focus();
-      } else if (this.form.get('bbDown')?.invalid) {
-        this.bbDownInput.nativeElement.focus();
-      } else if (this.form.get('location')?.invalid) {
-        this.locationInput.nativeElement.focus();
-      } else if (this.form.get('bbLr')?.invalid) {
-        this.bbLrInput.nativeElement.focus();
-      }
-    } else {
-      this.createCheckout();
-    }
-  }
-
-  createCheckout(): void {
-    if (!this.guestApprovedGDPR) {
-      this.snackbar.open('Guest needs to approve storing their data.', 'Okay', {
-        duration: 10000,
-      });
-      return;
-    }
-
-    this.isLoading = true;
-    this.service
-      .create({
-        room: this.form.get('room')?.value,
-        roomReady: this.form.get('roomReady')?.value,
-        name: this.form.get('name')?.value,
-        arrivalTime: new Date(this.form.get('arrivalTime')?.value),
-        bags: this.form.get('bags')?.value,
-        tagNr: this.form.get('tagNr')?.value,
-        bbLr: this.form.get('bbLr')?.value,
-        bbDown: this.form.get('bbDown')?.value,
-        location: this.form.get('location')?.value,
-        comments:
-          this.form.get('comments')?.value.toString().length > 1
-            ? this.form.get('comments')?.value
-            : '-',
-        luggageType: LuggageType.CHECKOUT,
-      })
-      .subscribe({
-        next: () => {
-          this.snackbar.open('Check In luggage item created!', 'Thanks', {
-            duration: 5000,
-          });
-          document.location.reload();
-          this.dialog.closeAll();
-          this.isLoading = false;
-        },
-        error: (err: HttpErrorResponse) => {
-          console.error(err);
-          this.snackbar.open('Failed to create, please try again.', 'Okay', {
-            duration: 10000,
-          });
-          this.isLoading = false;
-        },
-      });
-  }
-}
